requirements.txt
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cssselect==1.1.0
cymem==2.0.6
dateparser==1.1.1
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
Flask==2.0.3
gensim==4.1.2
GoogleNews==1.6.0
idna==3.3
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.0
joblib==1.1.0
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.1
murmurhash==1.0.6
newspaper3k==0.2.8
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathy==0.6.1
Pillow==9.0.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
requests-file==1.5.1
scipy==1.8.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.3
spacy-legacy==3.0.9
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.15
tinysegmenter==0.3
tldextract==3.2.0
tqdm==4.63.1
typer==0.4.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.1
urllib3==1.26.9
wasabi==0.9.0
Werkzeug==2.0.3
